+++ smrt/src/main/java/com/smart/dto/ColisRequestDTO.java
@@ -4,6 +4,8 @@ import jakarta.validation.Valid;
 import jakarta.validation.constraints.NotBlank;
 import jakarta.validation.constraints.NotNull;
 import lombok.Data;
+
+import java.math.BigDecimal;
 import java.util.List;
 
 @Data
@@ -12,7 +14,7 @@ public class ColisRequestDTO {
     private String description;
     
     @NotNull(message = "Le poids est requis")
-    private Double poids;
+    private BigDecimal poids;
     
     private String priorite; // HAUTE, MOYENNE, BASSE
