+++ smrt/src/main/java/com/smart/dto/ProduitDTO.java
@@ -2,11 +2,13 @@ package com.smart.dto;
 
 import lombok.Data;
 
+import java.math.BigDecimal;
+
 @Data
 public class ProduitDTO {
     private String id;
     private String nom;
     private String categorie;
-    private Double poids;
+    private BigDecimal poids;
     private Double prix;
 }
