+++ smrt/src/main/java/com/smart/entity/ColisProduit.java
@@ -5,6 +5,7 @@ import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
 
+import java.math.BigDecimal;
 import java.time.LocalDateTime;
 
 @Data
@@ -27,6 +28,6 @@ public class ColisProduit {
     private Produit produit;
 
     private Integer quantite;
-    private Double prix; // Prix au moment de l'ajout
+    private BigDecimal prix; // Prix au moment de l'ajout
     private LocalDateTime dateAjout;
 }
