+++ smrt/src/main/java/com/smart/entity/Produit.java
@@ -6,6 +6,7 @@ import jakarta.persistence.PrePersist;
 import lombok.AllArgsConstructor;
 import lombok.Data;
 import lombok.NoArgsConstructor;
+import java.math.BigDecimal;
 import java.util.UUID;
 
 @Data
@@ -19,8 +20,8 @@ public class Produit {
 
     private String nom;
     private String categorie;
-    private Double poids;
-    private Double prix;
+    private BigDecimal poids;
+    private BigDecimal prix;
 
     @PrePersist
     public void prePersist() {
